backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Settings
pydantic-settings==2.6.1

# Database
motor==3.6.0              # async MongoDB driver
pymongo==4.10.1

# Auth
authlib==1.3.2            # Google OAuth2
httpx==0.27.2             # async HTTP (authlib transport)
python-jose[cryptography]==3.3.0  # JWT sign/verify
passlib[bcrypt]==1.7.4

# Task queue
celery==5.4.0
redis==5.2.0
flower==2.0.1             # Celery monitoring UI

# Storage (S3-compatible)
boto3==1.35.74
botocore==1.35.74

# Vector search
qdrant-client==1.12.1

# Push notifications
pywebpush==2.0.0
py-vapid==1.9.0

# GCP
google-cloud-compute==1.19.3
google-auth==2.36.0

# PDF processing
pdfplumber==0.11.4
Pillow==11.0.0           # OG image generation

# Utilities
python-slugify==8.0.4
jinja2==3.1.4
aiofiles==24.1.0

# Scraper
feedparser==6.0.11          # RSS/Atom feed parsing
trafilatura==1.12.2         # article text extraction
beautifulsoup4==4.12.3      # HTML parsing
lxml==5.3.0                 # fast HTML/XML parser (bs4 backend)
